Initial commit
This is the first version, and it is super insecure.
Because this is not yet public, none of it is run behind HTTPS, and
there is no actual redirect URI set in oauth2.

Handling login and sessions, along with setting up HTTPS and a redirect
URI are the next big priorities.