Setup project to be run on a server
Previously the project required one to manually copy and enter an
offline token because it was not expecting to have access to a
RedirectURL.
Now it expects to have a RedirectURL, and is setup accordingly.

This commit removes a lot of cruft and creates two routes that properly
handle logging in - one that redirects to an OAuth2 endpoint and another
which handles being the OAuth2 RedirectURL endpoint.

Also, tokens are now persisted using an in memory session library.