Begin setting up a real database, use sqlite3
This is the start of using a real database.
It sets up many of the queries as prepared statements, though they are
not finalized.

Unfortunately, this change also uses a library which requires cgo.
That means it can no longer be easily cross-compiled for a different
system.
Eventually some kind of build tool might be nice, but building it on the
server is probably sufficiently fine for now.